Add files via upload
You recently offered samples of three products to every customer that entered your business. You then asked which product customers thought was the best. You recorded these preferences for Product A, Product B, and Product C into a dataset. Conduct a goodness of fit test to investigate this.

Do customers indicate a preference for one of the three products?